Refactor mla code from decode.py to mla.py
